Add involvement-api file and rename api.js (Cherrypick )
