+++ src/modules/displayData.js
@@ -1,9 +1,6 @@
-import getRecipes from './api.js';
+const container = document.getElementById('container');
 
-const container = document.querySelector('.grid-container');
-
-const displayCards = (recipes, nroLikes = 0) => {
-  
+const displayCards = (recipes) => {
   for (let i = 0; i < recipes.length; i += 1) {
     const card = document.createElement('ul');
     card.innerHTML = `<li><img src= ${recipes[i].thumbnail_url}></li>
